move generated sprite file to public folder
having this live in `addon/components` is potentially confusing as it isn't actually a component exposed by the addon
