+++ ember-flight-icons/public/icons/flight-icon-sprite.js
