+++ flight-icons/scripts/@types/ConfigData.d.ts
@@ -8,6 +8,5 @@ export type ConfigData = {
     mainFolder: string,
     tempFolder: string,
     emberPublicFolder: string,
-    emberAddonFolder: string,
     syncConcurrency: number,
 };
